[FLINK-6720] Rename java8 javaApiOperators package to api.java.operators
This closes #3999.
